refactor: Repository 인터페이스를 올바른 패키지로 이동
