+++ src/main/java/roomescape/repository/JdbcThemeRepository.java
@@ -7,7 +7,6 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.Theme;
-import roomescape.domain.ThemeRepository;
 
 @Repository
 public class JdbcThemeRepository implements ThemeRepository {
