+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -1,7 +1,8 @@
-package roomescape.domain;
+package roomescape.repository;
 
 import java.time.LocalDate;
 import java.util.List;
+import roomescape.domain.Reservation;
 
 public interface ReservationRepository {
