+++ src/main/java/roomescape/repository/ReservationTimeRepository.java
@@ -1,7 +1,8 @@
-package roomescape.domain;
+package roomescape.repository;
 
 import java.time.LocalTime;
 import java.util.List;
+import roomescape.domain.ReservationTime;
 
 public interface ReservationTimeRepository {
