+++ src/main/java/roomescape/repository/ThemeRepository.java
@@ -1,6 +1,7 @@
-package roomescape.domain;
+package roomescape.repository;
 
 import java.util.List;
+import roomescape.domain.Theme;
 
 public interface ThemeRepository {
